odic-esa-pipeline/requirements.txt
# ODIC ESA Pipeline Dependencies
# Python 3.11+

# LLM - OpenAI GPT-4o
openai>=1.12.0

# PDF Processing
PyPDF2>=3.0.0
pdfplumber>=0.10.0
reportlab>=4.0.0

# FTP/SFTP
paramiko>=3.4.0

# Config
PyYAML>=6.0.0

# API (optional status endpoints)
fastapi>=0.109.0
uvicorn>=0.27.0

# Async support
aiofiles>=23.2.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0

# Document conversion
python-docx>=1.1.0
Pillow>=10.0.0
PyMuPDF>=1.23.0

# Utilities
python-dotenv>=1.0.0
